Add internal docs for loader
